Delete keyboard messages in group
chat_id is now derived from chat in which message with keyboard was originally sent in. Meaning the bot will now respond to callbacks in the group to which the message was originally sent to and can delete it there. It no longer tries to delete it in the respondents personal chat.